requirements.txt
flask
flask-cors
python-dotenv
pygithub
beautifulsoup4
requests
retry
